start a pig farm
